Exercise-Snowflake-27-Time-Travel/requirements.txt
snowflake-connector-python==3.4.0
numpy==1.26.4
pandas==2.1.4
pyarrow==10.0.1
faker==19.6.2
python-dotenv==1.0.0
